% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_masshunter_csv.R
\name{parse_masshunter_csv}
\alias{parse_masshunter_csv}
\title{Read .csv files exported by Agilent's MassHunter software into an R data frame.}
\usage{
parse_masshunter_csv(filename)
}
\arguments{
\item{filename}{A string with the path of the relevant .csv file.}
}
\value{
A data frame with three columns.
}
\description{
Read .csv files exported by Agilent's MassHunter software into an R data frame.
}
