% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_masshunter_csv.R
\name{read_masshunter_csv}
\alias{read_masshunter_csv}
\title{Read a .csv file exported from MassHunter and assigns data blocks}
\usage{
read_masshunter_csv(filename)
}
\arguments{
\item{filename}{A string with the path of the relevant .csv file.}
}
\value{
A data frame with three columns.
}
\description{
Read a .csv file exported from MassHunter and assigns data blocks
}
